+++ core/src/main/java/cucumber/runtime/formatter/JUnitFormatter.java
@@ -12,6 +12,7 @@ import cucumber.api.event.TestStepFinished;
 import cucumber.api.formatter.Formatter;
 import cucumber.api.formatter.StrictAware;
 import cucumber.runtime.CucumberException;
+import cucumber.runtime.Utils;
 import cucumber.runtime.io.URLOutputStream;
 import cucumber.runtime.io.UTF8OutputStreamWriter;
 import org.w3c.dom.Document;
@@ -26,7 +27,6 @@ import javax.xml.transform.TransformerException;
 import javax.xml.transform.TransformerFactory;
 import javax.xml.transform.dom.DOMSource;
 import javax.xml.transform.stream.StreamResult;
-
 import java.io.Closeable;
 import java.io.IOException;
 import java.io.PrintWriter;
@@ -142,9 +142,6 @@ final class JUnitFormatter implements Formatter, StrictAware {
             rootElement.setAttribute("failures", String.valueOf(rootElement.getElementsByTagName("failure").getLength()));
             rootElement.setAttribute("skipped", String.valueOf(rootElement.getElementsByTagName("skipped").getLength()));
             rootElement.setAttribute("time", sumTimes(rootElement.getElementsByTagName("testcase")));
-            if (rootElement.getElementsByTagName("testcase").getLength() == 0) {
-                addDummyTestCase(); // to avoid failed Jenkins jobs
-            }
             TransformerFactory transfac = TransformerFactory.newInstance();
             Transformer trans = transfac.newTransformer();
             trans.setOutputProperty(OutputKeys.INDENT, "yes");
@@ -155,16 +152,6 @@ final class JUnitFormatter implements Formatter, StrictAware {
         } catch (TransformerException e) {
             throw new CucumberException("Error while transforming.", e);
         }
-    }
-
-    private void addDummyTestCase() {
-        Element dummy = doc.createElement("testcase");
-        dummy.setAttribute("classname", "dummy");
-        dummy.setAttribute("name", "dummy");
-        rootElement.appendChild(dummy);
-        Element skipped = doc.createElement("skipped");
-        skipped.setAttribute("message", "No features found");
-        dummy.appendChild(skipped);
     }
 
     private String sumTimes(NodeList testCaseNodes) {
@@ -230,16 +217,12 @@ final class JUnitFormatter implements Formatter, StrictAware {
         private String calculateElementName(cucumber.api.TestCase testCase) {
             String testCaseName = testCase.getName();
             if (testCaseName.equals(previousTestCaseName)) {
-                return testCaseName + (includesBlank(testCaseName) ? " " : "_") + ++exampleNumber;
+                return Utils.getUniqueTestNameForScenarioExample(testCaseName, ++exampleNumber);
             } else {
                 previousTestCaseName = testCase.getName();
                 exampleNumber = 1;
                 return testCaseName;
             }
-        }
-
-        private boolean includesBlank(String testCaseName) {
-            return testCaseName.indexOf(' ') != -1;
         }
 
         public void addTestCaseElement(Document doc, Element tc, Result result) {