+++ junit/src/main/java/cucumber/runtime/junit/JUnitOptions.java
@@ -18,7 +18,7 @@ public class JUnitOptions {
     /**
      * Create a new instance from a list of options, for example:
      * <p/>
-     * <pre<{@code Arrays.asList("--allow-started-ignored", "--filename-compatible-names");}</pre>
+     * <pre<{@code Arrays.asList("--filename-compatible-names", "--step-notifications");}</pre>
      *
      * @param argv the arguments
      */
@@ -34,10 +34,6 @@ public class JUnitOptions {
             if (arg.equals("--help") || arg.equals("-h")) {
                 printOptions();
                 System.exit(0);
-            } else if (arg.equals("--no-allow-started-ignored") || arg.equals("--allow-started-ignored")) {
-                System.err.println("WARNING: Found tags option '" + arg + "'. " +
-                        "--allow-started-ignored has no effect, testStarted is always fired before a test is started." +
-                        "--allow-started-ignored will be removed from the next release of Cucumber-JVM");
             } else if (arg.equals("--no-filename-compatible-names") || arg.equals("--filename-compatible-names")) {
                 filenameCompatibleNames = !arg.startsWith("--no-");
             } else if (arg.equals("--no-step-notifications") || arg.equals("--step-notifications")) {