+++ spring/src/main/java/cucumber/runtime/java/spring/SpringFactory.java
@@ -42,9 +42,6 @@ import java.util.Stack;
  * or @{@link BootstrapWith}. This step definition can also be annotated
  * with @{@link org.springframework.test.context.web.WebAppConfiguration}
  * or @{@link org.springframework.test.annotation.DirtiesContext} annotation.
- * <p>
- * If more that one step definition class has such an annotation, the annotations must be equal on
- * the different step definition. <b>Deprecation warning:</b> Annotating multiple step definitions is deprecated.
  * </li>
  * <li>If no step definition class with @ContextConfiguration or @ContextHierarchy
  * is found, it will try to load cucumber.xml from the classpath.</li>
@@ -58,8 +55,7 @@ import java.util.Stack;
  * the @Component stereotype an exception will be thrown
  * </li>
  * <li>
- * If more that one step definition class is used to configure the spring context, the annotations must be equal
- * on the different step definition. Please note that doing so is deprecated.
+ * If more that one step definition class is used to configure the spring context an exception will be thrown.
  * </li>
  * </ul>
  */
@@ -70,7 +66,6 @@ public class SpringFactory implements ObjectFactory {
 
     private final Collection<Class<?>> stepClasses = new HashSet<Class<?>>();
     private Class<?> stepClassWithSpringContext = null;
-    private boolean deprecationWarningIssued = false;
 
     public SpringFactory() {
     }
@@ -80,11 +75,12 @@ public class SpringFactory implements ObjectFactory {
         if (!stepClasses.contains(stepClass)) {
             checkNoComponentAnnotations(stepClass);
             if (dependsOnSpringContext(stepClass)) {
-                if (stepClassWithSpringContext == null) {
-                    stepClassWithSpringContext = stepClass;
-                } else {
-                    checkAnnotationsEqual(stepClassWithSpringContext, stepClass);
+                if (stepClassWithSpringContext != null) {
+                    throw new CucumberException(String.format("" +
+                        "Glue class %1$s and %2$s both attempt to configure the spring context. Please ensure only one " +
+                        "glue class configures the spring context", stepClass, stepClassWithSpringContext));
                 }
+                stepClassWithSpringContext = stepClass;
             }
             stepClasses.add(stepClass);
         }
@@ -126,41 +122,6 @@ public class SpringFactory implements ObjectFactory {
                 }
             }
 
-        }
-        return false;
-    }
-
-
-    private void checkAnnotationsEqual(Class<?> stepClassWithSpringContext, Class<?> stepClass) {
-        if (!deprecationWarningIssued) {
-            deprecationWarningIssued = true;
-            System.err.println(String.format("" +
-                    "WARNING: Having more than one glue class that configures " +
-                    "the spring context is deprecated. Found both %1$s and %2$s " +
-                    "that attempt to configure the spring context.",
-                    stepClass, stepClassWithSpringContext));
-        }
-        Annotation[] annotations1 = stepClassWithSpringContext.getAnnotations();
-        Annotation[] annotations2 = stepClass.getAnnotations();
-        if (annotations1.length != annotations2.length) {
-            throw new CucumberException("Annotations differs on glue classes found: " +
-                    stepClassWithSpringContext.getName() + ", " +
-                    stepClass.getName());
-        }
-        for (Annotation annotation : annotations1) {
-            if (!isAnnotationInArray(annotation, annotations2)) {
-                throw new CucumberException("Annotations differs on glue classes found: " +
-                        stepClassWithSpringContext.getName() + ", " +
-                        stepClass.getName());
-            }
-        }
-    }
-
-    private boolean isAnnotationInArray(Annotation annotation, Annotation[] annotations) {
-        for (Annotation annotationFromArray: annotations) {
-            if (annotation.equals(annotationFromArray)) {
-                return true;
-            }
         }
         return false;
     }